+++ packages/core/test/core/hmr-perf/incremental-graph/dependency-reparse/long-chain.ts
@@ -1,0 +1,48 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../../helpers';
+
+describe('Issue #1223 HMR perf - incremental graph dependency reparsing - long chain', () => {
+  test('phase2b: 10단계 체인에서 leaf 변경 시 reparsedModules=1', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-1223-phase2b-'));
+    const N = 10;
+    for (let i = 0; i < N - 1; i++) {
+      writeFileSync(
+        join(dir, `m${i}.ts`),
+        `import { v${i + 1} } from "./m${i + 1}"; export const v${i} = v${i + 1} + 1;`,
+      );
+    }
+    writeFileSync(join(dir, `m${N - 1}.ts`), `export const v${N - 1} = 1;`);
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const { promise: rebuildP, resolve: rebuildDone } = Promise.withResolvers<any>();
+    const handle = watch({
+      entryPoints: [join(dir, 'm0.ts')],
+      devMode: true,
+      collectModuleCodes: true,
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuildDone(event);
+      },
+    });
+    await readyP;
+    await new Promise((r) => setTimeout(r, 100));
+
+    writeFileSync(join(dir, `m${N - 1}.ts`), `export const v${N - 1} = 999;`);
+    const event = await rebuildP;
+    handle.stop();
+    rmSync(dir, { recursive: true });
+
+    expect(event.reparsedModules).toBe(1);
+  }, 15000);
+});
