+++ packages/core/test/core/hmr-perf/incremental-graph/dependency-reparse/middle-module.ts
@@ -1,0 +1,43 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../../helpers';
+
+describe('Issue #1223 HMR perf - incremental graph dependency reparsing - middle module', () => {
+  test('phase2c: 체인 중간(b)만 변경 — 상위(a)/하위(c) 캐시 유지', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-1223-phase2c-'));
+    writeFileSync(join(dir, 'a.ts'), 'import { b } from "./b"; export const a = b + 1;');
+    writeFileSync(join(dir, 'b.ts'), 'import { c } from "./c"; export const b = c + 1;');
+    writeFileSync(join(dir, 'c.ts'), 'export const c = 1;');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const { promise: rebuildP, resolve: rebuildDone } = Promise.withResolvers<any>();
+    const handle = watch({
+      entryPoints: [join(dir, 'a.ts')],
+      devMode: true,
+      collectModuleCodes: true,
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuildDone(event);
+      },
+    });
+    await readyP;
+    await new Promise((r) => setTimeout(r, 100));
+
+    writeFileSync(join(dir, 'b.ts'), 'import { c } from "./c"; export const b = c + 42;');
+    const event = await rebuildP;
+    handle.stop();
+    rmSync(dir, { recursive: true });
+
+    expect(event.reparsedModules).toBe(1);
+  }, 10000);
+});
